Refactor detais to dapp_details
