--- client/modules/admin/actions/index.js
@@ -1,2 +1,0 @@
-export default {
-};
