--- client/modules/admin/index.js
@@ -1,10 +1,0 @@
-import actions from './actions';
-import routes from './routes.jsx';
-
-export default {
-  routes,
-  actions,
-  load(context) {
-    
-  }
-};
