--- client/modules/admin/routes.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-import {mount} from "react-mounter";
-import MainLayout from "/client/modules/core/components/main_layout.jsx";
-
-export default function (injectDeps, {FlowRouter}) {
-  const MainLayoutCtx = injectDeps(MainLayout);
-
-  FlowRouter.route('/login', {
-    name: 'login',
-    action() {
-      console.log('admin');
-      /* mount(MainLayoutCtx, {
-       content: () => (< />)
-       });*/
-    }
-  });
-};
