+++ client/modules/dapps/components/dapp_details.jsx
