+++ client/modules/dapps/containers/dapp_details.js
@@ -1,6 +1,6 @@
 import {useDeps, composeAll, composeWithTracker, compose} from 'mantra-core';
 
-import Details from '../components/details.jsx';
+import DappDetails from '../components/dapp_details.jsx';
 
 export const composer = ({context}, onData) => {
   const {Meteor, Collections} = context();
