+++ client/modules/dapps/routes.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {mount} from "react-mounter";
 import MainLayout from "/client/modules/core/components/main_layout.jsx";
-import Details from "/client/modules/dapps/containers/details";
+import DappDetails from "/client/modules/dapps/containers/dapp_details";
 export default function (injectDeps, {FlowRouter}) {
   const MainLayoutCtx = injectDeps(MainLayout);
 
@@ -10,7 +10,7 @@ export default function (injectDeps, {FlowRouter}) {
     action() {
       console.log('details page route');
       mount(MainLayoutCtx, {
-        content: () => (<Details/>)
+        content: () => (<DappDetails/>)
       });
     }
   });
